move isIgnoredPath to dedicated Path.ts
