+++ packages/@expo/fingerprint/src/hash/Hash.ts
@@ -1,7 +1,6 @@
 import { createHash } from 'crypto';
 import { createReadStream } from 'fs';
 import fs from 'fs/promises';
-import minimatch from 'minimatch';
 import pLimit from 'p-limit';
 import path from 'path';
 
@@ -13,6 +12,7 @@ import type {
   HashSourceContents,
   NormalizedOptions,
 } from '../Fingerprint.types';
+import { isIgnoredPath } from '../utils/Path';
 import { profile } from '../utils/Profile';
 
 /**
@@ -129,31 +129,6 @@ export async function createFileHashResultsAsync(
       });
     });
   });
-}
-
-/**
- * Indicate the given `filePath` should be excluded by `ignorePaths`
- */
-export function isIgnoredPath(
-  filePath: string,
-  ignorePaths: string[],
-  minimatchOptions: minimatch.IOptions = { dot: true }
-): boolean {
-  const minimatchObjs = ignorePaths.map(
-    (ignorePath) => new minimatch.Minimatch(ignorePath, minimatchOptions)
-  );
-
-  let result = false;
-  for (const minimatchObj of minimatchObjs) {
-    const currMatch = minimatchObj.match(filePath);
-    if (minimatchObj.negate && result && !currMatch) {
-      // Special handler for negate (!pattern).
-      // As long as previous match result is true and not matched from the current negate pattern, we should early return.
-      return false;
-    }
-    result ||= currMatch;
-  }
-  return result;
 }
 
 /**
